introduce batch type-specific options as key=value pairs
Some batch systems may have additional options that don't fit neatly into
the frame we already have, e.g., SGE has a parallel environment (pe) that
needs to be specified to get resources.  Changed the "--batch-options"
(which was used to add preamble options) to "--batch-preamble" and then
added "--batch-options" to take key=value pairs that can be used for the
additional options.